cotctg/requirements.txt
aenum==2.0.8
backports.shutil-get-terminal-size==1.0.0
certifi==2017.7.27.1
chardet==3.0.4
dbf==0.96.8
decorator==4.1.2
Django==1.11.4
django-allauth==0.33.0
django-rest-auth==0.9.1
djangorestframework==3.6.3
enum34==1.1.6
fpdf==1.7.2
httplib2==0.7.3
idna==2.6
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
oauthlib==2.0.2
olefile==0.44
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.7.11
PySimpleSOAP==1.8.8
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
